requirements.txt
flask==2.3.3
python-dotenv==1.0.0
plexapi==4.15.4
pandas==2.1.1
numpy==1.26.0
openai==1.3.5
anthropic==0.5.0
chromadb==0.4.18
requests==2.31.0